Complete ship model tests and add skeleton of other model tests
